requirements.txt
playwright==1.48.0
beautifulsoup4==4.12.3
chromadb==0.5.5
google-generativeai==0.8.3
python-dotenv==1.0.1
google-api-core==2.22.0
flask